requirements.txt
coverage==6.3.2
Django==3.2.15
djangorestframework==3.12.4
drf-yasg
model-bakery==1.7.0
psycopg2==2.8.5
